If-Else Homework Commit

+++ src/Mathematics_of_two_numbers.java
@@ -4,15 +4,15 @@ public class Mathematics_of_two_numbers {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.println("Enter First Number");
-        int a = sc.nextInt();
+        double a = sc.nextInt();
         System.out.println("Enter Second Number");
-        int b = sc.nextInt();
+       double b = sc.nextInt();
         System.out.println("Enter Mathematic any Symbol (Ex:+,-,/,*)");
         char symbol = sc.next().charAt(0);
-        int addition = a+b;
-        int subtraction = a-b;
-        int division = a/b;
-        int multiplication = a*b;
+        double addition = a+b;
+        double subtraction = a-b;
+        double division = a/b;
+        double multiplication = a*b;
         if(symbol=='+'){
             System.out.println("Adition of above numbers is " + addition);
         }else if(symbol=='-'){
